# code/sonara_back_end/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sonara API",
    "description": "\nAPI do projeto **Sonara** — plataforma de conexão entre artistas e organizadores de eventos.\n\n## Autenticação\nAlgumas rotas exigem **Bearer Token** (JWT). Faça login em `POST /v1/sonara/usuario/login` e use o token retornado no header:\n`Authorization: Bearer <token>`\n\n## Fluxo principal\n1. Cadastre um usuário (artista ou organizador)\n2. Faça login para obter o token\n3. Crie eventos (organizador) ou candidate-se a eventos (artista)\n4. Gerencie candidaturas, contra-propostas e convites\n        ",
    "version": "1.0.10",
    "contact": {
      "name": "Davi de Almeida Santos"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Insira: Bearer <token>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/sonara/artista/": {
      "get": {
        "tags": [
          "Artista"
        ],
        "summary": "Listar todos os artistas",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum artista encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Artista"
        ],
        "summary": "Cadastrar artista",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Artista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Artista criado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/artista/{id}": {
      "get": {
        "tags": [
          "Artista"
        ],
        "summary": "Buscar artista por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Artista encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "ID inválido"
          },
          "404": {
            "description": "Artista não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Artista"
        ],
        "summary": "Atualizar artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do artista"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Artista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artista atualizado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Artista não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Artista"
        ],
        "summary": "Excluir artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Artista excluído"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Artista não encontrado"
          }
        }
      }
    },
...
    "/v1/sonara/evento/": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Listar todos os eventos",
        "description": "Retorna todos os eventos cadastrados com fotos, endereço e organizador.",
        "responses": {
          "200": {
            "description": "Lista de eventos retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum evento encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Evento"
        ],
        "summary": "Criar evento",
        "description": "Cria evento com endereço e vínculo ao organizador.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Evento criado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos obrigatórios ausentes ou inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/evento/{id}": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Buscar evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Evento"
        ],
        "summary": "Atualizar evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento atualizado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento"
        ],
        "summary": "Excluir evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento excluído com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      }
    },
...
    "/v1/sonara/eventoArtista/listar_por_evento/{idEvento}": {
      "get": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Listar inscrições por evento",
        "description": "Retorna todos os artistas inscritos em um evento específico. Usado pelo organizador.",
        "parameters": [
          {
            "name": "idEvento",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de inscrições retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhuma inscrição encontrada para este evento"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/minhas_candidaturas/{artistaId}": {
      "get": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Listar candidaturas do artista",
        "description": "Retorna todas as candidaturas de um artista. Usado pelo próprio artista.",
        "parameters": [
          {
            "name": "artistaId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de candidaturas retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhuma candidatura encontrada"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/aceitarConvite/{id}": {
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Artista aceita convite do organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Convite aceito com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/recusarConvite/{id}": {
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Artista recusa convite do organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Convite recusado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/": {
      "get": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Listar todos os vínculos evento-artista",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum vínculo encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Criar vínculo evento-artista manualmente",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoArtista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Vínculo criado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/candidatar": {
      "post": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Artista se candidata a um evento",
        "description": "O artista se inscreve em um evento informando o cachê esperado, motivação e descrição.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Candidatura"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Candidatura realizada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos obrigatórios ausentes ou inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/aprovar/{id}": {
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Organizador aprova artista no evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Artista aprovado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/reprovar/{id}": {
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Organizador reprova artista no evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Artista reprovado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/contraProposta/{id}": {
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Organizador envia contra proposta de cachê ao artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ContraProposta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contra proposta enviada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/aceitarContraProposta/{id}": {
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Artista aceita contra proposta do organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Contra proposta aceita com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/recusarContraProposta/{id}": {
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Artista recusa contra proposta do organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Contra proposta recusada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      }
    },
    "/v1/sonara/eventoArtista/{id}": {
      "get": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Buscar vínculo evento-artista por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Vínculo encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "ID inválido"
          },
          "404": {
            "description": "Não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Atualizar vínculo evento-artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoArtista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vínculo atualizado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento x Artista"
        ],
        "summary": "Excluir vínculo evento-artista",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do vínculo evento-artista"
          }
        ],
        "responses": {
          "200": {
            "description": "Vínculo excluído com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Vínculo não encontrado"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Artista": {
      "type": "object",
      "properties": {
        "nome_artistico": {
          "type": "string",
          "example": "DJ João"
        },
        "usuario_id": {
          "type": "number",
          "example": 1
        },
        "descricao": {
          "type": "string",
          "example": "DJ com 5 anos de experiência"
        }
      },
      "required": [
        "nome_artistico",
        "usuario_id",
        "descricao"
      ]
    },
...
  }
}